Fix for ubsan warning when multiplying a negative value with an unsigned int

+++ lib/Remotery.c
@@ -3328,7 +3328,7 @@ static rmtU32 rotl32(rmtU32 x, rmtS8 r)
 static rmtU32 getblock32(const rmtU32* p, int i)
 {
     rmtU32 result;
-    const rmtU8* src = ((const rmtU8*)p) + i * sizeof(rmtU32);
+    const rmtU8* src = ((const rmtU8*)p) + i * (int)sizeof(rmtU32);
     memcpy(&result, src, sizeof(result));
     return result;
 }
